Added HTTP Response Tagging API
The CacheViewEventListener is modified to use Taggers and a ResponseCacheConfigurator.

The ResponseCacheConfigurator centralizes cache configuration handling for a Response:

```
$configurator->makePublic($response);
$configurator->addTags($response, ['location-2', 'content-345']);
```

Any Value Object may have a Tagger. A Tagger can add tags to a Response,
by mapping properties from the Value it handles:

    $tagger->tag($configurator, $response, $content);

To add the tags to the Response, Taggers are given the ResponseCacheConfigurator.
In addition to Taggers that map values from the value objects, other taggers may
extract a value object, and delegate its tagging to another Tagger. One example
is ContentView, that is delegated to ContentInfoTagger by the ContentValueViewTagger.

This commit also integrates phpspec. The objects defined here are all covered by their
own specs, that tries to describe the behaviour that can be expected from them.